Updated source code to java 7. Reformatted to consistently use 4 spaces as indentation depth. Fixed javadoc.

+++ maltcms-datastructures/src/main/java/maltcms/datastructures/peak/normalization/TicAreaNormalizer.java
@@ -60,7 +60,7 @@ public class TicAreaNormalizer implements IPeakNormalizer {
 
     @Override
     public double getNormalizationFactor(IFileFragment fragment, int peakIndex) {
-        return 1.0d / getCache().get(fragment).doubleValue();
+        return 1.0d / getCache().get(fragment);
     }
 
     @Override
